update: 삭제전 확인창 추가

+++ pages/_category/_postId/index.vue
@@ -148,18 +148,21 @@ export default {
       this.$router.push({ name: 'write', params: { postId } })
     },
     async deletePost(postId) {
-      // 게시글을 삭제
-      await this.$store
-        .dispatch('deletePost', postId)
-        .catch((e) => {
-          // 서버에서 에러가 발생했을때
-          this.isOpen = true
-          this.dialogType = 'error'
-          this.dialogMessage = e.message
-        })
-        .finally(() => {
-          this.$store.commit('SET_LOADING', false)
-        })
+      const isDelete = confirm('정말로 삭제하시겠습니까?')
+      if (isDelete) {
+        // 게시글을 삭제
+        await this.$store
+          .dispatch('deletePost', postId)
+          .catch((e) => {
+            // 서버에서 에러가 발생했을때
+            this.isOpen = true
+            this.dialogType = 'error'
+            this.dialogMessage = e.message
+          })
+          .finally(() => {
+            this.$store.commit('SET_LOADING', false)
+          })
+      }
     },
     async writeComment() {
       // 작성한 댓글의 정보
